services/cubic/requirements.txt
pycryptodome
redis